+++ src/AcuityAtomicSwapSell.t.sol
@@ -15,8 +15,8 @@ contract AcuityAtomicSwapSellTest is DSTest {
     }
 
     function testCreateOrder() public {
-        uint price = 5;
-        uint value = 50;
+        uint64 price = 5;
+        uint64 value = 50;
         bytes32 orderId = acuityAtomicSwapSell.createOrder{value: value}(price);
         assertEq(address(acuityAtomicSwapSell).balance, value);
         assertEq(acuityAtomicSwapSell.getOrderValue(address(this), price), value);
@@ -24,124 +24,124 @@ contract AcuityAtomicSwapSellTest is DSTest {
     }
 
     function testControlLockSellNotBigEnough() public {
-        uint price = 5;
-        uint value = 50;
+        uint64 price = 5;
+        uint64 value = 50;
         acuityAtomicSwapSell.createOrder{value: value}(price);
-        acuityAtomicSwapSell.lockSell(price, hex"1234", block.timestamp + 1000, value);
+        acuityAtomicSwapSell.lockSell(price, hex"1234", uint64(block.timestamp) + 1000, value);
     }
 
     function testFailLockSellNotBigEnough() public {
-        uint price = 5;
-        uint value = 50;
+        uint64 price = 5;
+        uint64 value = 50;
         acuityAtomicSwapSell.createOrder{value: value}(price);
-        acuityAtomicSwapSell.lockSell(price, hex"1234", block.timestamp + 1000, value + 10);
+        acuityAtomicSwapSell.lockSell(price, hex"1234", uint64(block.timestamp) + 1000, value + 10);
     }
 
     function testLockSell() public {
-        uint price = 5;
+        uint64 price = 5;
         bytes32 orderId = acuityAtomicSwapSell.createOrder{value: 50}(price);
 
         bytes32 secret = hex"1234";
         bytes32 hashedSecret = keccak256(abi.encodePacked(secret));
-        uint timeout = block.timestamp + 1000;
-        uint value = 10;
+        uint64 timeout = uint64(block.timestamp) + 1000;
+        uint64 value = 10;
         acuityAtomicSwapSell.lockSell(price, hashedSecret, timeout, value);
         assertEq(acuityAtomicSwapSell.getOrderValue(orderId), 40);
-        (bytes32 lockOrderId, uint lockValue, uint lockTimeout) = acuityAtomicSwapSell.getSellLock(hashedSecret);
+        (bytes32 lockOrderId, uint64 lockValue, uint64 lockTimeout) = acuityAtomicSwapSell.getSellLock(hashedSecret);
         assertEq(lockOrderId, orderId);
         assertEq(lockValue, value);
         assertEq(lockTimeout, timeout);
     }
 
     function testControlUnlockSellTimedOut() public {
-        uint price = 5;
+        uint64 price = 5;
         acuityAtomicSwapSell.createOrder{value: 50}(price);
         bytes32 secret = hex"1234";
         bytes32 hashedSecret = keccak256(abi.encodePacked(secret));
-        acuityAtomicSwapSell.lockSell(price, hashedSecret, block.timestamp + 1000, 40);
+        acuityAtomicSwapSell.lockSell(price, hashedSecret, uint64(block.timestamp) + 1000, 40);
         acuityAtomicSwapSell.unlockSell(secret);
     }
 
     function testFailUnlockSellTimedOut() public {
-        uint price = 5;
+        uint64 price = 5;
         acuityAtomicSwapSell.createOrder{value: 50}(price);
         bytes32 secret = hex"1234";
         bytes32 hashedSecret = keccak256(abi.encodePacked(secret));
-        acuityAtomicSwapSell.lockSell(price, hashedSecret, block.timestamp, 40);
+        acuityAtomicSwapSell.lockSell(price, hashedSecret, uint64(block.timestamp), 40);
         acuityAtomicSwapSell.unlockSell(secret);
     }
 
     function testUnlockSell() public {
-        uint price = 5;
+        uint64 price = 5;
         bytes32 orderId = acuityAtomicSwapSell.createOrder{value: 50}(price);
         bytes32 secret = hex"1234";
         bytes32 hashedSecret = keccak256(abi.encodePacked(secret));
-        uint timeout = block.timestamp + 1000;
-        uint value = 10;
+        uint64 timeout = uint64(block.timestamp) + 1000;
+        uint64 value = 10;
 
         acuityAtomicSwapSell.lockSell(price, hashedSecret, timeout, value);
-        (bytes32 lockOrderId, uint lockValue, uint lockTimeout) = acuityAtomicSwapSell.getSellLock(hashedSecret);
+        (bytes32 lockOrderId, uint64 lockValue, uint64 lockTimeout) = acuityAtomicSwapSell.getSellLock(hashedSecret);
         assertEq(lockOrderId, orderId);
         assertEq(lockValue, value);
         assertEq(lockTimeout, timeout);
 
         assertEq(address(acuityAtomicSwapSell).balance, 50);
-        uint startBalance = address(this).balance;
+        uint64 startBalance = uint64(address(this).balance);
         acuityAtomicSwapSell.unlockSell(secret);
         (lockOrderId, lockValue, lockTimeout) = acuityAtomicSwapSell.getSellLock(hashedSecret);
         assertEq(lockOrderId, 0);
         assertEq(lockValue, 0);
         assertEq(lockTimeout, 0);
         assertEq(address(acuityAtomicSwapSell).balance, 40);
-        assertEq(address(this).balance, startBalance + 10);
+        assertEq(uint64(address(this).balance), startBalance + 10);
     }
 
     function testControlTimeoutSellWrongOrderId() public {
-        uint price = 5;
+        uint64 price = 5;
         acuityAtomicSwapSell.createOrder{value: 50}(price);
         bytes32 secret = hex"1234";
         bytes32 hashedSecret = keccak256(abi.encodePacked(secret));
-        acuityAtomicSwapSell.lockSell(price, hashedSecret, block.timestamp, 10);
+        acuityAtomicSwapSell.lockSell(price, hashedSecret, uint64(block.timestamp), 10);
         acuityAtomicSwapSell.timeoutSell(price, hashedSecret);
     }
 
     function testFailTimeoutSellWrongOrderId() public {
-        uint price = 5;
+        uint64 price = 5;
         acuityAtomicSwapSell.createOrder{value: 50}(price);
         bytes32 secret = hex"1234";
         bytes32 hashedSecret = keccak256(abi.encodePacked(secret));
-        acuityAtomicSwapSell.lockSell(price, hashedSecret, block.timestamp, 10);
+        acuityAtomicSwapSell.lockSell(price, hashedSecret, uint64(block.timestamp), 10);
         acuityAtomicSwapSell.timeoutSell(price - 1, hashedSecret);
     }
 
     function testControlTimeoutSellNotTimedOut() public {
-        uint price = 5;
+        uint64 price = 5;
         acuityAtomicSwapSell.createOrder{value: 50}(price);
         bytes32 secret = hex"1234";
         bytes32 hashedSecret = keccak256(abi.encodePacked(secret));
-        acuityAtomicSwapSell.lockSell(price, hashedSecret, block.timestamp, 10);
+        acuityAtomicSwapSell.lockSell(price, hashedSecret, uint64(block.timestamp), 10);
         acuityAtomicSwapSell.timeoutSell(price, hashedSecret);
     }
 
     function testFailTimeoutSellNotTimedOut() public {
-        uint price = 5;
+        uint64 price = 5;
         acuityAtomicSwapSell.createOrder{value: 50}(price);
         bytes32 secret = hex"1234";
         bytes32 hashedSecret = keccak256(abi.encodePacked(secret));
-        acuityAtomicSwapSell.lockSell(price, hashedSecret, block.timestamp + 1000, 10);
+        acuityAtomicSwapSell.lockSell(price, hashedSecret, uint64(block.timestamp) + 1000, 10);
         acuityAtomicSwapSell.timeoutSell(price, hashedSecret);
     }
 
     function testTimeoutSell() public {
-        uint price = 5;
+        uint64 price = 5;
         bytes32 orderId = acuityAtomicSwapSell.createOrder{value: 50}(price);
         bytes32 secret = hex"1234";
         bytes32 hashedSecret = keccak256(abi.encodePacked(secret));
-        uint timeout = block.timestamp;
-        uint value = 10;
+        uint64 timeout = uint64(block.timestamp);
+        uint64 value = 10;
 
         acuityAtomicSwapSell.lockSell(price, hashedSecret, timeout, value);
-        (bytes32 lockOrderId, uint lockValue, uint lockTimeout) = acuityAtomicSwapSell.getSellLock(hashedSecret);
+        (bytes32 lockOrderId, uint64 lockValue, uint64 lockTimeout) = acuityAtomicSwapSell.getSellLock(hashedSecret);
         assertEq(lockOrderId, orderId);
         assertEq(lockValue, value);
         assertEq(lockTimeout, timeout);
